+++ CallCenter/src/main/java/com/redcard/telephone/service/TelephoneRecordManager.java
@@ -54,7 +54,7 @@ public class TelephoneRecordManager extends GenericPageHQLQuery<TelephoneRecord>
 		telephoneRecord.setFldCallEndTime(calllog.getHangUpTime());
 		telephoneRecord.setFldCallLong(calllog.getTalkDuration());
 		telephoneRecord.setFldCallDate(calllog.getInboundCallTime());
-		Talklog talkLog = talklogDao.findByCallId(Integer.valueOf(callId));
+		Talklog talkLog = talklogDao.findByCallId(Long.valueOf(callId));
 		telephoneRecord.setFldRecordFilePath(talkLog.getIispath());
 		telephoneRecordDao.save(telephoneRecord);
 	}
